requirements.txt
numpy
pandas 
matplotlib
seaborn 
scikit-learn
streamlit 
python-dotenv
google-generativeai
